Initial recipe structur eproposal
