+++ lib/conjuror.options.js
@@ -10,6 +10,14 @@ argv.option({
 });
 
 argv.option({
+  name: 'recipe',
+  short: 'r',
+  type: 'string',
+  description: 'Points at what recipe to use.',
+  example: "'conjuror.js --recipe=value' or 'conjuror.js -r create-clients'. Recipe can either be a predefined recipe, or a custom one."
+});
+
+argv.option({
   name: 'format',
   short: 'f',
   type: 'csv,string',
