+++ lib/conjuror.recipes.js
@@ -1,0 +1,10 @@
+var Recipes = {
+  "create-clients": {
+
+    "name": "Create Clients File",
+    "description": "create a clients file",
+    "path": "../recipes/createClients.recipe.json"
+  }
+};
+
+module.exports = Recipes;
